% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_nas.R
\name{clean_data}
\alias{clean_data}
\title{Remove NA values and return clean data}
\usage{
clean_data(filepath)
}
\arguments{
\item{filepath}{\itemize{
\item the directory to data
}}
}
\value{
Data with no NA values present
}
\description{
Remove NA values and return clean data
}
