% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{make_boxplot}
\alias{make_boxplot}
\title{Create a boxplot with the variable species_id and weight}
\usage{
make_boxplot(filepath)
}
\arguments{
\item{filepath}{\itemize{
\item the directory to data
}}
}
\value{
Boxplot of species_id versus weight
}
\description{
Create a boxplot with the variable species_id and weight
}
